requirements.txt
langchain-core
langgraph
langchain
streamlit
matplotlib
pandas
numpy
chromadb
langgraph==1.0.4
typing-extensions
langchain==0.1.20
langchain-community==0.0.38
faiss-cpu==1.11.0.post1
sentence-transformers==2.7.0
pypdf==4.3.1
python-dotenv==1.0.1
pandas==2.2.3
matplotlib==3.9.2
pydantic>=2.7.1,<3.0
langchain_mistralai
nltk
rouge-score
streamlit
transformers
torch
sentencepiece
accelerate
bitsandbytes
langchain_huggingface
langchain_core
huggingface-hub==0.26.2
transformers==4.46.2
huggingface_hub[inference]
chromadb
graphviz
python-dotenv
langchain-text-splitters
